requirements.txt
llama-index-core==0.10.50
llama-index-readers-file==0.1.25
llama-index-llms-nvidia==0.1.3
llama-index-embeddings-nvidia==0.1.4
llama-index-postprocessor-nvidia-rerank==0.1.2
azure-cognitiveservices-speech
openai
flask
flask-cors
soundex
tensorflow
gevent

